==> 03 Scripts/01.IC Data import and descriptive analysis.ipynb <==
# 01. Importing Libraries

# import libraries
import pandas as pd
import numpy as np
import os

# 02. Importing Data

orders.csv

path = r'/Users/bazillaimran/Documents/02-2023 Instacart Basket Analysis'
path

df = pd.read_csv(os.path.join(path,'02 Data','Original Data', 'orders.csv'), index_col = False)

vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)

df.head()

df.shape

products.csv

df_prods = pd.read_csv(os.path.join(path,'02 Data','Original Data','products.csv'), index_col = False)

df_prods.head(20)

df_prods.tail(35)

df_prods.columns

df_prods.shape

df_prods.describe()

What’s the maximum value of the “aisle_id” column? 134.000000

df_prods.dtypes

What’s the data type of the “department_id” column? 64 bit Integer
