==> 03 Scripts/02. Data Wrangling and Subsetting .ipynb <==
# 01. Importing Libraries and Data

# importing libraries
import pandas as pd
import numpy as np
import os

path = r'/Users/bazillaimran/Documents/02-2023 Instacart Basket Analysis'

df_ords = pd.read_csv(os.path.join(path,'02 Data','Original Data', 'orders.csv'), index_col = False)

df_prods = pd.read_csv(os.path.join(path,'02 Data','Original Data', 'products.csv'), index_col = False)

df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col=False)

# 02. Wrangling Procedures

df_ords.columns

# Dropping eval_set column from orders.csv
df_ords = df_ords.drop(columns = ['eval_set'])

df_ords.head()

df_ords['days_since_prior_order'].value_counts(dropna = False)

# Renaming order_dow
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

df_ords.head()

# Changing Variable Type
df_ords['order_id'] = df_ords['order_id'].astype('str')

df_ords['order_id'].dtype

# Transposing df_dep
df_dep_t=df_dep.T

df_dep_t

# Getting rid of '0' with an index
df_dep_t

# Getting rid of "0" with an index
df_dep_t.reset_index()

new_header = df_dep_t.iloc[0]

new_header

df_dep_t_new = df_dep_t[1:]

df_dep_t_new

df_dep_t_new.columns = new_header

df_dep_t_new

# 02. Changing another variable

# Changing another Variable
df_ords['user_id'] = df_ords['user_id'].astype('str')

df_ords['user_id'].dtype

# 03. Renaming variable in df.ords

df_ords.rename(columns = {'order_hour_of_day' : 'time_of_day'}, inplace = True)

df_ords.head()

# 04. Frequency of placing orders by the hour

Method # 1

df_ords.groupby('time_of_day').size()

Highest Frequency Hour: 10  

Method # 2

hourly_frequency = df_ords['time_of_day'].value_counts()

print(hourly_frequency)

Hour 10 has the Highest Frequency

# 05. Determining the meaning of value 4 in department_id

data_dict = df_dep_t_new.to_dict('index')

print(data_dict.get('4'))

Department_id of 4 stands for produce

# 06. Breakfast Item Sales

data_dict

Department_id: 14 for Breakfast

# Creating a subset for breakfast (Method # 1)
df_breakfast = df_prods[df_prods['department_id']==14]

df_breakfast.head()

# Creating a subset for breakfast (Method # 2)
df_breakfast_2 = df_prods.loc[df_prods['department_id'] == 14]

loc = looks for a column name

# Creating a subset for breakfast (Method # 3)
df_Breakfast_3 = df_prods.loc[df_prods['department_id'].isin([14])]

isin([]) = Telling the loc function to look into a list ex: (isin([14,17,18])

# 07. Subsets: alcohol, deli, Beverages, meat/seafood

data_dict

department_id of intereset: 5,20,7,12

df_dinner_parties = df_prods.loc[df_prods['department_id'].isin([5,20,7,12])]

df_dinner_parties

# 08. Total counts of rows for df_dinner_parties

df_dinner_parties.shape

# 09. info. on user_id 1

df_ords_user_id_1 = df_ords.loc[df_ords['user_id'] == "1"]

df_ords_user_id_1.shape

df_ords_user_id_1.info()

# 10. Basic stats on user_id 1

df_ords_user_id_1.describe()

# 11. Exporting prepared data

df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

df_dep_t_new.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'departments_wrangled.csv'))
