==> 03 Scripts/03.Data Consistency Checks.ipynb <==
#import libraries
import pandas as pd
import numpy as np
import os

# assigning path of the dataframes to 'path' variable
path=r'/Users/bazillaimran/Documents/02-2023 Instacart Basket Analysis/02 Data'

# import products dataframe from original folder
df_prods=pd.read_csv(os.path.join(path,'Original Data','products.csv'),index_col=False)

# import orders_wrangled dataframe from prepared folder
df_ords=pd.read_csv(os.path.join(path,'Prepared Data','orders_wrangled.csv'),index_col=False)

# 01. Step 1: Consistency Checks on df_prods

df_prods.describe()

# Address Mixed data-types

# Create a dataframe
df_test = pd.DataFrame()

# Create a mixed type column
df_test['mix'] = ['a','b',1,True]

df_test.head()

# The function for checking whether a dataframe contains any mixed-type columns is as follows:

for col in df_test.columns.tolist():
  weird = (df_test[[col]].map(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)


df_test['mix'] = df_test['mix'].astype('str')

# Missing Values

# Find missing values in products dataframe
# The output shows there are 16 missing values in product_name colum

df_prods.isnull().sum()

# to view 16 observations with missing values, you can create a subset of the dataframe 
# containing only the values in question

df_nan=df_prods[df_prods['product_name'].isnull()==True]

df_nan

# df.describe()

# df['column with missings'].fillna(mean value, inplace=True)

# df_prods.median()

# df['column with missings'].fillna(median value, inplace=True)

# Check the number of rows and columns before clean up
df_prods.shape

# Method 1: extract only the observations from df_prods where product_name column is not blank and store it in a new dataframe
# df_prods_clean

df_prods_clean= df_prods[df_prods['product_name'].isnull() == False]

# Check the number of rows and columns in the new dataframe with extracted records after clean up

df_prods_clean.shape

# Method 2: drop all missing values from df_prods useing:

# df_prods.dropna(inplace=True) -- For all columns in df_prods

# If you wanted to drop only the NaNs from a particular column.  Inplace=True with replace the existing dataframe with new.

# df_prods.dropna(subset='product_name',inplace=True)

# Duplicates

# Finding Duplicates -- extracting duplicate from df_prods_clean and store it in dataframe df_dups
df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

# Addressing Duplicates -- 1. Check the current number of rows and columns
df_prods_clean.shape

# 2. Create a new dataframe that doesn’t include the duplicates (df_dups)
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

# export df_prods_clean_no_dups to prepared data folder as a new file name: products_checked.csv

df_prods_clean_no_dups.to_csv(os.path.join(path,'Prepared Data', 'products_checked.csv'))

# 02 Step: Consistency Checks on df_ords 

# Run the df.describe() function on your df_ords dataframe
df_ords.describe()

The "day_sine_prior_order" column has a right-skewed distribution, evident from a higher mean (11.102) compared to the median (7.000), indicating a tail towards higher values. The large standard deviation (9.186) suggests significant variability. The notable difference between the mean and median implies potential outliers in the upper range.

The order_number column shows a right-skewed distribution with significant variability, possibly influenced by higher order numbers.

Conclusion: The presence of outliers and the right-skewed nature of both order_numbers and days_since_prior_order suggest a consistent pattern. It's possible that certain customers are placing more orders, leading to higher order_numbers that appear as outliers. Similarly, some customers may not shop frequently on Instacart, resulting in higher days_since_prior_order counts.

# Check for mixed data-types columns
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].map(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)
else :
    print ("None of the columns have mix data")

# Check for missing values
df_ords.isnull().sum()

df_ords.shape

# creat a new data frame df_nan for null values of column day_since_prior_order
df_ords_nan = df_ords[df_ords['days_since_prior_order'].isnull() == True]

df_ords_nan

df_ords_nan['order_number'].value_counts(dropna = False)

df_ords_nan['user_id'].value_counts(dropna = False)

Answer 5: Considering the substantial number of missing values (206209), it's crucial to investigate the reason before resorting to imputation or deletion, as these approaches may have detrimental effects. Upon analysis, it appears that all 206209 missing values are associated with order_number 1, aligning with the expectation that first orders wouldn't have a prior order date. Further, when counting user_ids with NaN values for prior dates, we find that 206209 users have ordered only once, consistent with the NaN count.

# Addressing Missing Values

Answer 6: Hence, retaining the NaN values in this column is reasonable as it preserves the meaningful context, indicating the absence of prior orders for users making their first order.

# Checking for Duplicate Values 
df_ords_dups = df_ords[df_ords.duplicated()]

df_ords_dups

Answer 7/8: No Duplicates found

df_ords.shape

# Export df_ords as orders_checked
df_ords.to_csv(os.path.join(path,'Prepared Data', 'orders_checked.csv'))
