==> 03 Scripts/04.Combining & Exporting Data 1.ipynb <==
# 01. Importing Data

# Importing libraries

import pandas as pd
import numpy as np
import os

# Defining path
path = r'/Users/bazillaimran/Documents/02-2023 Instacart Basket Analysis'

# Importing data set orders_checked
df_ords = pd.read_csv(os.path.join(path,'02 Data', 'Prepared Data','orders_checked.csv'), index_col = False)

# Importing data set orders_prior
df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders_products_prior.csv'), index_col = False)

# Check the output df_ords
df_ords.head()

df_ords_prior.head()

# Check shape of df_ords
df_ords.shape

# Check shape of df_ords_prior
df_ords_prior.shape

# create new merged dataframe -- inner join
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

df_merged_large.head()

df_merged_large['_merge'].value_counts()

# Export merged dataframe as pickle
df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))
