==> 03 Scripts/05.Combining and Exporting Data 2.ipynb <==
# import libraries
import pandas as pd
import numpy as np
import os

Question 3: Importing Dataframes

path = r'/Users/bazillaimran/Documents/02-2023 Instacart Basket Analysis'

# Import orders_products_combined
df_ords_prods_combined = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

# Import df_prods
df_prods =  pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'), index_col = False)

Question 4: Check the shape of the imported dataframe

# Check shape of orders_products_combined
df_ords_prods_combined.shape

# Check shape of df_prods
df_prods.shape

Question 5: Determine a suitable way to combine the 2 dataframes

df_ords_prods_combined.head()

df_prods.head()

# Drop columns in df_ords_prods_combined (not necessary) 
df_ords_prods_combined = df_ords_prods_combined.drop(columns = ['Unnamed: 0', 'Unnamed: 0.1', '_merge'])

# Drop unnamed columns in df_prods (not necessary)
df_prods = df_prods.drop(columns = ['Unnamed: 0'])

# Merge by column product_id since it's present in both dataframes 
df_ords_prods_merged = df_prods.merge(df_ords_prods_combined, on = 'product_id', indicator = True)

Question 6: Confirm the results of the merge using the merge flag.

# checking df_ords_prods_merged dataframe
df_ords_prods_merged.head()

#Checking the values in the _merge column using merge flag
df_ords_prods_merged['_merge'].value_counts()

#Checking the merged shape
df_ords_prods_merged.shape

Question 7: Export the newly created dataframe as orders_products_merged 

#Exporting the newly merged dataframe; Using pickle format
df_ords_prods_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged.pkl'))
