==> 03 Scripts/06.Deriving New Variables.ipynb <==
# Importing Data 

import pandas as pd
import pandas as np
import os

# Defing project folder path
path = r'/Users/bazillaimran/Documents/02-2023 Instacart Basket Analysis/02 Data'

path

df_ords_prods_merge = pd.read_pickle(os.path.join(path,'Prepared Data','orders_products_merged.pkl'))

# 1a: Creating "price_label" column

# Creating a subset of first one million rows
df = df_ords_prods_merge[:1000000]

df.shape

def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

df['price_range'] = df.apply(price_label, axis=1)

df['price_range'].value_counts(dropna = False)

df['prices'].max()

# Use loc() method on df subset to apply price range function.
df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df['price_range_loc'].value_counts(dropna = False)

# Use loc() method on entire ords_prods_merge dataframe to apply price range function.
df_ords_prods_merge.loc[df_ords_prods_merge['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_ords_prods_merge.loc[(df_ords_prods_merge['prices'] <= 15) & (df_ords_prods_merge['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

df_ords_prods_merge.loc[df_ords_prods_merge['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_ords_prods_merge['price_range_loc'].value_counts(dropna = False)

# 1b. Create "busiest_day" column

# Change "order_dow" column name to "order_day_of_week."
df_ords_prods_merge.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

# Find frequency of orders_day_of_week.

df_ords_prods_merge['orders_day_of_week'].value_counts(dropna = False)

# Use for-loop to determine whether orders are on "busiest day" (0 = Saturday), "least busy" (4 = Wednesday), or "regularly busy" (other days of the week).

result = []

for value in df_ords_prods_merge["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

# Create new column in ords_prods_merge to view results in context.

df_ords_prods_merge['busiest_day'] = result

df_ords_prods_merge['busiest_day'].value_counts(dropna = False)

# 2. Update"busiest day" column to "busiest days" to identify order day of the week as "Busiest days," "Slowest days," or "Regularly busy."

# Create column “busiest_days" from "orders_day_of_week" to identify two busiest days (0, 1), slowest days (4, 3), or regularly busy (all other days).

result_2 = []

for value in df_ords_prods_merge["orders_day_of_week"]:
  if value == 0 or value == 1:
    result_2.append("Busiest days")
  elif value == 4 or value == 3:
    result_2.append("Slowest days")
  else:
    result_2.append("Regularly busy")

result_2

# 3. Check values of new "busiest days" column for accuracy with observations in markdown format.

# Create new 'busiest_days' column in ords_prods_merge to view results in context.

df_ords_prods_merge['busiest_days'] = result_2

df_ords_prods_merge['busiest_days'].value_counts(dropna = False)

# Check output of ords_prods_merge with new "busiest days" column.

df_ords_prods_merge.head()

Both columns for the busiest_day(s) are consistent with the total count generated for the orders_day_of_week parameter. For instance, in the plural column, 3 and 4 are assigned to the slowest days category. Therefore, when we add the counts for both 3 and 4 (3840534 + 3783802), we get the total count for the slowest days, which is 7624336 which is consistant with the count for the busiest_days column. 

# 4. Create new column "busiest_period_of_day" to identify time periods “Most orders,” “Average orders,” and “Fewest orders.”

# Change "time_of_day" column name to "order_hour_of_day."

df_ords_prods_merge.rename(columns = {'time_of_day' : 'order_hour_of_day'}, inplace = True)

# Check value counts in "order_hour_of_day" column.

df_ords_prods_merge['order_hour_of_day'].value_counts()

Split into: 
"Most orders" = 10, 11, 14, 15, 13, 12, 16, and 9; 
"Fewest orders" = 23, 6, 0, 1, 5, 2, 4, and 3; 
"Average orders" = all remaining values in order_hour_of_day.

# Create column “busiest_period_of_day" from "order_hour_of_day" to identify "Most orders", "Fewest orders", or "Average orders".

result_3 = []

for value in df_ords_prods_merge["order_hour_of_day"]:
  if value in [10,11,14,15,13,12,16,9]:
    result_3.append("Most orders")
  elif value in [23,6,0,1,5,2,4,3]:
    result_3.append("Fewest orders")
  else:
    result_3.append("Average orders")

result_3

# Create new column "busiest_period_of_day" in ords_prods_merge.

df_ords_prods_merge['busiest_period_of_day'] = result_3

# 5. Print frequency of "busiest_period_of_day" column.

# Print value counts in "busiest_period_of_day" column.

df_ords_prods_merge['busiest_period_of_day'].value_counts(dropna = False)

# 7. Export dataframe as a pickle file to “Prepared Data” folder.

#Exporting the dataframe; Using pickle format
df_ords_prods_merge.to_pickle(os.path.join(path,'Prepared Data', 'orders_products_merged_derived.pkl'))
