==> 03 Scripts/08.Data Visualization with Python Part 1.ipynb <==
# Import Libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

path = r'/Users/bazillaimran/Documents/02-2023 Instacart Basket Analysis/02 Data'

df_customers = pd.read_csv(os.path.join(path,'Original Data','customers.csv'))

# Data Wrangling: rename columns with logical names and drop unwanted columns 

df_customers.head()

df_customers.shape

# rename columns
df_customers.rename(columns = {'Surnam':'Last Name','fam_status':'family_status','STATE':'State','n_dependants':'dependants'}, inplace = True)

df_customers.head()

All columns contain valuable information, there is no necessity to eliminate any of them.

# Data quality and consistency checks: missing values, duplicates, and mixed-data conversion

df_customers.describe()

Income column: If the mean and median are close in value, it suggests that the income distribution is relatively symmetric. In other words, the data is evenly distributed around the center.

# Address mixed-data types

df_customers.info()

No mixed datatype in each column. However, a more appropriate data type for the date_joined column would be datetime data type.

# Change data type for object to datetime for date_joined column
df_customers['date_joined'] = pd.to_datetime(df_customers['date_joined'])

df_customers.info()

# Check for duplicates
# extracting duplicate from df_customers and store it in dataframe df_dups
df_dups = df_customers[df_customers.duplicated()]

df_dups

No duplicates found.

# Check for missing values
df_customers.isnull().sum()

The output shows there are 11259 missing values in the First Name column.

# Viewing missing values: create a subset of the dataframe containing only the missing values
df_nan=df_customers[df_customers['First Name'].isnull()==True]

df_nan

Since the missing information is only in the "First Name" column for 11,259 rows, it's better to keep them. Deleting these rows could make our data less reliable, so it's a good idea to leave them as they are. This way, we keep more data and make sure our analysis is based on as much information as possible. However, we can enhance the comprehensibility of our data by replacing these missing values with 'Unknown'.

# replace NaN with Unknown
df_customers['First Name'].fillna('Unknown', inplace = True)

df_customers.isnull().sum()

df_customers['First Name'].value_counts()

# Combining customer data with prepared data


# Load prepared data
df_all=pd.read_pickle(os.path.join(path, 'Prepared Data', 'ords_prods_grouped.pkl'))

# Combine df_cutomer with df_all
# Check shape of df_customer
df_customers.shape

# Check shape of df_all
df_all.shape

# check data type and determine a suitable way to combine the 2 dataframes
df_all.info()

df_customers.info()

# Drop _merge columns in df_all (not necessary) 
df_all = df_all.drop(columns = ['_merge'])

Merge the to dataframes by using the user_id columns (same data type of int64). 

# Merge by column user_id since it's present in both dataframes 
ords_prods_customers_merged = df_all.merge(df_customers, on = 'user_id', indicator = True)

# Confirm the results of the merge using the merge flag.
ords_prods_customers_merged.head()

#Checking the values in the _merge column using merge flag
ords_prods_customers_merged['_merge'].value_counts()

ords_prods_customers_merged.shape

# Exporting New Dataframe as a pickle file

ords_prods_customers_merged = ords_prods_customers_merged.to_pickle(os.path.join(path,'Prepared Data', 'ords_prods_customers_merged.pkl'))
